+++ src/crawlee/playwright_crawler/types.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
-from crawlee.basic_crawler.types import BasicCrawlingContext
+from crawlee.basic_crawler.types import BasicCrawlingContext, EnqueueLinksFunction
 
 if TYPE_CHECKING:
     from playwright.async_api import Page
@@ -14,3 +14,4 @@ class PlaywrightCrawlingContext(BasicCrawlingContext):
     """Crawling context used by PlaywrightSoupCrawler."""
 
     page: Page
+    enqueue_links: EnqueueLinksFunction